requirements.txt
python-telegram-bot
telegram
typing
python-dotenv
PyPDF2
openai
pyperclip